Add a stack.yaml file
I needed this when building cabal.exe for MinGHC (see:
https://github.com/fpco/minghc/issues/69). I thought others might find
this useful, especially for the use case of bootstrapping cabal
development.

I went the direction of no curation and explicit packages in the
extra-deps, as I thought that would make more sense for a project low on
the dependency chain, but this could certainly be tweaked to use LTS
Haskell or Stackage Nightly instead.