Change a few InstallPlan functions to use all dependencies, including setup deps
When setup deps were introduced "Add ComponentSetup to ComponentDeps", we
kept all uses to be the same old deps, without changing anything by using
those new setup deps. In several places in InstallPlan it makes sense to
use all deps, including setup deps. In fact for InstallPlan we should always
use all deps except in one place: in checking plan validity where we check
for dependency inconsistencies. That's the one place where we look only at
"normal" non-setup deps to check that we're using only one version of each
package.

So for example, with this patch the freeze command now includes setup deps
in the frozen set. The other case that should change is that failures of
deps of packages used only for setup scripts should now be handled properly
because we will look at setup dep edges when looking for things that depend
on a failed package.