Use the new package db utils to simplify createInternalPackageDB
And fix up the db path for UHC. UHC cannot just register anywhere, like
compilers that have a hc-pkg can. It has to be special locations. Now
that registerPackage no longer takes the inplace :: Bool arg, UHC's impl
of registerPackage has to get the dir from the PackageDbStack without
knowing if it's implace or not. So the correct inplace location has to
be set earlier for the inplace package db, which is what this does.